renamed pages to screens and added search from jio saavn feature
