+++ packages/media/index.ts
@@ -1,5 +1,7 @@
 import * as Youtube from './src/Plugins/Youtube';
+import * as JioSaavn from './src/Plugins/JioSaavn';
 
 export {
-    Youtube
+    Youtube,
+    JioSaavn
 }
