+++ packages/mobile/src/screens/home/Main.tsx
@@ -2,7 +2,6 @@ import React, { useRef } from 'react';
 import { ScrollView, View } from 'react-native';
 
 import { useScrollToTop } from '@react-navigation/native';
-import { useSelector } from 'react-redux';
 import { NetNotify } from '../../components/NetNotify';
 import { RecentContainer } from '../../containers/RecentContainer';
 import { MostPlayedContainer } from '../../containers/MostPlayedContainer';
@@ -10,8 +9,8 @@ import OnlineContainer from '../../containers/OnlineContainer';
 import { Screen } from '../../components/Screen';
 import { ShortCutContainer } from '../../containers/ShortcutContainer';
 import OnlineSongsContainer from '../../containers/OnlineSongsContainer';
-import YoutubeSongsContainer from '../../containers/YoutubeSongsContainer';
-import { JioSaavnContainer } from '../../containers/JioSaavnContainer';
+import YoutubeSongsContainer from './components/YoutubeSongsList';
+import { JioSaavnContainer } from './components/JioSaavnList';
 
 const Divider = () => <View style={{ marginVertical: 8 }} />;
