+++ packages/mobile/src/screens/home/components/DiagnoseDialog.tsx
