+++ packages/mobile/src/screens/player/index.tsx
